refactor(presentation): UI 패키지 화면별로 변경
