+++ app/src/main/res/layout/activity_choice_seat.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray_100"
-    tools:context=".presentation.activity.ChoiceSeatActivity">
+    tools:context=".presentation.choiceSeat.ChoiceSeatActivity">
 
     <ScrollView
         android:layout_width="match_parent"
