+++ app/src/main/res/layout/activity_completed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.activity.CompletedActivity">
+    tools:context=".presentation.complete.CompleteActivity">
 
     <TextView
         android:id="@+id/textCompletedDescription"
